Update dependencies; fix fallout form pep440_rs change from usize => u64 (#588)

+++ rye/src/sources.rs
@@ -118,11 +118,7 @@ impl From<PythonVersion> for Version {
     fn from(value: PythonVersion) -> Self {
         Version {
             epoch: 0,
-            release: vec![
-                value.major as usize,
-                value.minor as usize,
-                value.patch as usize,
-            ],
+            release: vec![value.major as u64, value.minor as u64, value.patch as u64],
             pre: None,
             post: None,
             dev: None,
@@ -136,9 +132,9 @@ impl From<PythonVersionRequest> for Version {
         Version {
             epoch: 0,
             release: vec![
-                value.major as usize,
-                value.minor.unwrap_or_default() as usize,
-                value.patch.unwrap_or_default() as usize,
+                value.major as u64,
+                value.minor.unwrap_or_default() as u64,
+                value.patch.unwrap_or_default() as u64,
             ],
             pre: None,
             post: None,
